+++ .env
@@ -1,5 +1,6 @@
 API_PORT=8081
 API_CORS_ORIGIN=http://localhost:8081
+API_CORS_CREDENTIALS=true
 API_DB_NOMBRE=recibos
 API_DB_USUARIO=postgres
 API_DB_CONTRASENA=negativo10
